Back/OCR_Service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

numpy==1.26.4
pillow==10.3.0

paddleocr==2.7.3
paddlepaddle>=2.6.2

pdf2image==1.17.0
python-pptx==0.6.23

pymysql==1.1.0
requests==2.31.0